test: move LLMAgent registry construction tests to test_agent.py
Moves the subagents_registry (Ch9) and a2a_agents_registry (Ch10)
construction/duplicate-guard tests from test_task_handler.py to
test_agent.py — they exercise LLMAgent.__init__ directly and never
touch TaskHandler, matching the existing test_init/
test_init_raises_error_duplicated_tools pattern already in
test_agent.py. Dropped the now-unnecessary @pytest.mark.asyncio,
since none of these tests actually await anything.

Also moves the subagents tests for consistency, even though they
predate this PR and weren't the ones flagged.
