+++ tests/agent/test_agent.py
@@ -3,7 +3,9 @@ import contextlib
 from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
+from a2a.types import AgentCard, AgentInterface
 
+from llm_agents_from_scratch.a2a import A2AAgentSpec
 from llm_agents_from_scratch.agent import LLMAgent
 from llm_agents_from_scratch.base.llm import BaseLLM
 from llm_agents_from_scratch.base.tool import BaseTool
@@ -27,6 +29,7 @@ from llm_agents_from_scratch.skills.constants import (
     EXPLICIT_SKILL_ACTIVATION_TEMPLATE,
     EXPLICIT_SKILL_ACTIVATION_WITH_PROMPT_TEMPLATE,
 )
+from llm_agents_from_scratch.subagents import SubAgentSpec
 
 
 def test_init(mock_llm: BaseLLM) -> None:
@@ -458,3 +461,84 @@ async def test_run_approval_gate_rejection_does_not_increment_step_counter(
     # No run_step was ever called (get_next_step always returned TaskResult),
     # so the rejection must not have incremented step_counter.
     assert handler.step_counter == 0
+
+
+# ---------------------------------------------------------------------------
+# Subagents tests (Chapter 9)
+# ---------------------------------------------------------------------------
+
+
+def test_llm_agent_init_with_subagents(mock_llm: BaseLLM) -> None:
+    """Tests LLMAgent builds subagents_registry from the provided list."""
+    spec = SubAgentSpec(
+        name="researcher",
+        description="Looks things up.",
+        agent=LLMAgent(llm=mock_llm),
+    )
+    agent = LLMAgent(llm=mock_llm, subagents=[spec])
+
+    assert "researcher" in agent.subagents_registry
+    assert agent.subagents_registry["researcher"] is spec
+
+
+def test_llm_agent_init_no_subagents_defaults_to_empty_dict(
+    mock_llm: BaseLLM,
+) -> None:
+    """Tests LLMAgent.subagents_registry defaults to empty dict."""
+    agent = LLMAgent(llm=mock_llm)
+
+    assert agent.subagents_registry == {}
+
+
+def test_llm_agent_init_raises_on_duplicate_subagent_names(
+    mock_llm: BaseLLM,
+) -> None:
+    """Tests LLMAgent raises LLMAgentError on duplicate subagent names."""
+    spec = SubAgentSpec(
+        name="researcher",
+        description="Looks things up.",
+        agent=LLMAgent(llm=mock_llm),
+    )
+    with pytest.raises(LLMAgentError, match="duplicate"):
+        LLMAgent(llm=mock_llm, subagents=[spec, spec])
+
+
+# ---------------------------------------------------------------------------
+# A2A tests (Chapter 10)
+# ---------------------------------------------------------------------------
+
+
+def _a2a_spec(name: str) -> A2AAgentSpec:
+    card = AgentCard(
+        name=name,
+        description="A peer agent.",
+        supported_interfaces=[AgentInterface(url="http://peer:9999")],
+    )
+    return A2AAgentSpec.from_agent_card(agent_card=card)
+
+
+def test_llm_agent_init_with_a2a_agents(mock_llm: BaseLLM) -> None:
+    """Tests LLMAgent builds a2a_agents_registry from the provided list."""
+    spec = _a2a_spec("researcher")
+    agent = LLMAgent(llm=mock_llm, a2a_agents=[spec])
+
+    assert "researcher" in agent.a2a_agents_registry
+    assert agent.a2a_agents_registry["researcher"] is spec
+
+
+def test_llm_agent_init_no_a2a_agents_defaults_to_empty_dict(
+    mock_llm: BaseLLM,
+) -> None:
+    """Tests LLMAgent.a2a_agents_registry defaults to empty dict."""
+    agent = LLMAgent(llm=mock_llm)
+
+    assert agent.a2a_agents_registry == {}
+
+
+def test_llm_agent_init_raises_on_duplicate_a2a_agent_names(
+    mock_llm: BaseLLM,
+) -> None:
+    """Tests LLMAgent raises LLMAgentError on duplicate a2a_agent names."""
+    spec = _a2a_spec("researcher")
+    with pytest.raises(LLMAgentError, match="duplicate"):
+        LLMAgent(llm=mock_llm, a2a_agents=[spec, spec])
